Reworked draw wiki logic due changes in original code
