+++ src/index.css
@@ -239,12 +239,16 @@ svg text {
 .section h3 {
 	text-align: center;
 }
-#home_resources, #home_items, #home_planets {
+
+#home_resources,
+#home_items,
+#home_planets {
 	display: flex;
 	flex-wrap: wrap;
 	justify-content: space-between;
 }
-:is(#home_resources, #home_items, #home_planets) > li {
+
+:is(#home_resources, #home_items, #home_planets)>li {
 	width: 20rem;
 }
