+++ src/labels.json
@@ -26,5 +26,8 @@
 	"planet_at_core": {"en-US": "At core", "fr-FR": "Au noyau"},
 	"planet_atmospheric_resources": {"en-US": "Atmospheric resources", "fr-FR": "Ressources atmosphériques"},
 	"wiki_link": {"en-US": "Open article on Astroneer Wiki", "fr-FR": "Ouvrir l'article sur Astroneer Wiki"},
-	"wiki_category": {"en-US": "Category", "fr-FR": "Catégorie"}
+	"wiki_category": {"en-US": "Category", "fr-FR": "Catégorie"},
+	"resource": {"en-US": "Resource", "fr-FR": "Ressource"},
+	"item": {"en-US": "Item", "fr-FR": "Objet"},
+	"planet": {"en-US": "Planet", "fr-FR": "Planète"}
 }
